Moved all classes in one module
