--- dynatrace_event/stackstate_checks/dynatrace_event/dynatrace_exception.py
@@ -1,10 +1,0 @@
-# (C) StackState 2020
-# All rights reserved
-# Licensed under a 3-clause BSD style license (see LICENSE)
-
-
-class EventLimitReachedException(Exception):
-    """
-    Exception raised when maximum number of event reached
-    """
-    pass
